requirements.txt
# Core dependencies for Phase 1
google-api-python-client==2.106.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
pandas==2.1.3
numpy==1.24.3
pyyaml==6.0.1
python-dateutil==2.8.2

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# Development
black==23.11.0
flake8==6.1.0
isort==5.12.0

# Phase 2: Sentiment Analysis & Visualization
textblob==0.17.1
vaderSentiment==3.3.2
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0
wordcloud==1.9.2

# Phase 3: Web Dashboard
streamlit==1.28.2

# Future phases (for advanced features)
# scikit-learn==1.3.2
# spacy==3.7.2
# transformers==4.35.2
# openpyxl==3.1.2
# reportlab==4.0.7
